Module Lib.Server

exception Login_failure of string

val login : Yojson.Basic.t -> unit

val make_response : int64 option -> Opium_kernel.Rock.Response.t Lwt.t

val load_json : Opium.Std.Request.t -> 'a -> (Yojson.Basic.t -> 'a -> 'b) -> 'b Lwt.t

val load_json_login : Opium.Std.Request.t -> 'a -> (Yojson.Basic.t -> 'a -> int64 option) -> Opium_kernel.Rock.Response.t Lwt.t

load_json req ins_func inserter inserts the contents of load_json into the database

val json_of_user : Dbquery.user -> [> Ezjsonm.t ]

val json_of_group : Dbquery.group -> [> Ezjsonm.t ]

val json_of_restriction_id : string -> [> Ezjsonm.t ]

val is_member : int -> int -> bool

Checks to see if u_id is in g_id

val is_friend : int -> int -> bool

Checks to see if two users are friends, with f1 being the sender

val user_inserter : Yojson.Basic.t -> (string -> string -> string -> 'a) -> 'a

user_inserter json ins_func inserts the data representing a user into the database

val friend_inserter : Yojson.Basic.t -> (int -> int -> 'a) -> 'a

val rest_inserter : Yojson.Basic.t -> (int -> int -> 'a) -> 'a

val rest_indx_inserter : Yojson.Basic.t -> (string -> 'a) -> 'a

val group_info_inserter : Yojson.Basic.t -> (string -> int -> 'a) -> 'a

val group_inserter : Yojson.Basic.t -> (int -> int -> 'a) -> 'a

val group_host_user_inserter : Yojson.Basic.t -> (int -> int -> int -> 'a) -> 'a

val survey_inserter : Yojson.Basic.t -> (int -> int -> float -> float -> string -> int -> int -> 'a option) -> 'a option

val vote_status_inserter : Yojson.Basic.t -> (int -> int -> 'a) -> 'a

val vote_inserter : Yojson.Basic.t -> (int -> int -> int list -> 'a option) -> 'a option

val default : Opium.Std.App.builder

val get_list : Opium.Std.App.builder list

val post_list : Opium.Std.App.builder list

val app_builder : ('a -> 'a) list -> 'a -> 'a

val port : int

val start : unit -> unit